+++ otc/src/commands/mod.rs
@@ -1,3 +1,4 @@
+pub mod bench;
 pub mod lib;
 pub mod new;
 pub mod project;